+++ kernel/gc.c
@@ -4,7 +4,7 @@
 gc_t *gc = NULL;
 
 void destructor_gc() {
-    for (int i = 0; i < gc->number_objects; ++i) {
+    for (long long i = 0; i < gc->number_objects; ++i) {
         destructor_object(gc->objects[i]);
     }
     free(gc->objects);
@@ -21,13 +21,13 @@ void init_stack_frame() {
     gc->number_objects = 0;
 }
 
-int new_object(class_t *class) {
+long long new_object(class_t *class) {
     gc->objects = realloc(gc->objects, sizeof(object_t *));
 
     object_t *object = malloc(sizeof(object_t));
     object->class = class;
-    object->fields = malloc(sizeof(int) * class->number_fields);
+    object->fields = malloc(sizeof(long long) * class->number_fields);
 
     gc->objects[gc->number_objects] = object;
-    return (int) object;
+    return (long long) object;
 }
