+++ kernel/gc.h
@@ -10,8 +10,8 @@
 typedef struct gc gc_t;
 struct gc{
     object_t **objects;
-    int number_objects;
+    long long number_objects;
 };
 
-int new_object(class_t *class);
+long long new_object(class_t *class);
 #endif //VIRTUALMACHINE_GC_H
