+++ kernel/stack_frame.c
@@ -6,12 +6,12 @@ stack_frame_t *stack_frame = NULL;
 void destructor_stack_frame() {
     free(stack_frame->stack_frame);
 
-    for (int i = 0; i < stack_frame->number_frames; ++i) {
+    for (long long i = 0; i < stack_frame->number_frames; ++i) {
         destructor_frame(stack_frame->frames[i]);
     }
     free(stack_frame->frames);
 
-    for (int i = 0; i < stack_frame->number_classes; ++i) {
+    for (long long i = 0; i < stack_frame->number_classes; ++i) {
         destructor_class(stack_frame->classes[i]);
     }
     free(stack_frame->classes);
@@ -37,8 +37,8 @@ stack_frame_t *get_stack_frame() {
     return stack_frame;
 }
 
-frame_t *find_frame(int index_frame) {
-    for (int i = 0; i < stack_frame->number_frames; ++i) {
+frame_t *find_frame(long long index_frame) {
+    for (long long i = 0; i < stack_frame->number_frames; ++i) {
         if (stack_frame->frames[i]->index_frame == index_frame) {
             return stack_frame->frames[i];
         }
@@ -47,8 +47,8 @@ frame_t *find_frame(int index_frame) {
     return NULL;
 }
 
-class_t *find_class(int index_class){
-    for (int i = 0; i < stack_frame->number_frames; ++i) {
+class_t *find_class(long long index_class){
+    for (long long i = 0; i < stack_frame->number_frames; ++i) {
         if (stack_frame->classes[i]->index_class == index_class) {
             return stack_frame->classes[i];
         }
